Fix zoomed card zooming in and out during AI turn
I think I got it! Tricky one because there were two causes: 1) during render the cards are drawn in order then the zoomed card placed on top - the temporary order change could trigger a pointerout event, and 2) the code to move the zoomed card fully onto the screen was causing the card in hand to jump up to meet it temporarily during render.
Both of these are resolved and I have not seen the glitch again during testing (except when an accessed card is zoomed but that's intended behaviour not a glitch).